feat: Parent Stack Addition (PIQ-44)
It turns out that you need a blank Stack that the Git Sync configuration can be attached to.  It is a little odd how this was all done, and because of this oddity, doing this via multple stacks is barely better than ClickOps.

The only thing that makes this any better than ClickOps is the fact that you elimiate human error.